pyproject.toml
[tool.poetry]
name = "telegram-bot-python"
version = "0.1.0"
description = ""
authors = ["Hao Nguyen <hao@haonguyen.tech>"]
readme = "README.md"
packages = [{include = "telegram_bot_python"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.3.1"
python-dotenv = "^1.0.0"
pytz = "^2023.3"
requests = "^2.29.0"
python-telegram-bot = "^20.2"
mypy = "^1.2.0"
types-requests = "^2.28.11.17"
types-pytz = "^2023.3.0.0"
pylint = "^2.17.3"
pynvim = "^0.4.3"
black = "^23.3.0"
flake8 = "^6.0.0"
pyright = "^1.1.305"

[tool.black]
line-length = 100
target-version = ['py37']
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/foo.py    # exclude a file named foo.py in the root of the project
  | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''

[tool.pyright]
typeCheckingMode = "basic"
defineConstant = { DEBUG = true }
reportMissingImports = true
exclude = ["**/node_modules",
    "**/__pycache__",
    "src/experimental",
    "src/typestubs",
    ".mypy_cache",
    ".pytest_cache",
    "venv"
]
venvPath = "./venv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
